+++ src/main/java/absolutelyaya/ultracraft/data/LevelDataManager.java
@@ -29,8 +29,7 @@ public class LevelDataManager extends JsonDataLoader
 	public static final Identifier PLACEHOLDER_THUMB = new Identifier(Ultracraft.MOD_ID, "textures/level/placeholder.png");
 	private static final Gson GSON = new GsonBuilder().setPrettyPrinting().disableHtmlEscaping().create();
 	public static LevelDataManager Instance;
-	public static Map<Identifier, LevelData> levels = new HashMap<>();
-	public static Map<Identifier, LevelData> customLevels = new HashMap<>();
+	public static Map<Identifier, LevelData> levels = new HashMap<>(), customLevels = new HashMap<>();
 	
 	public LevelDataManager()
 	{
@@ -61,14 +60,13 @@ public class LevelDataManager extends JsonDataLoader
 	@Override
 	protected void apply(Map<Identifier, JsonElement> prepared, ResourceManager manager, Profiler profiler)
 	{
-		ImmutableMap.Builder<Identifier, LevelData> builtinBuilder = ImmutableMap.builder();
-		ImmutableMap.Builder<Identifier, LevelData> customBuilder = ImmutableMap.builder();
+		ImmutableMap.Builder<Identifier, LevelData> builtinBuilder = ImmutableMap.builder(), customBuilder = ImmutableMap.builder();
 		prepared.forEach((id, element) -> {
 			JsonObject json = element.getAsJsonObject();
 			boolean builtin = JsonHelper.getBoolean(json, "builtin", false);
 			if(!json.has("structure"))
 			{
-				Ultracraft.LOGGER.warn((builtin ? "Level '" : "Custom Level '") + id + "' does not have a structure parameter!");
+				Ultracraft.LOGGER.warn("{} '{}' does not have a structure parameter!", builtin ? "Level" : "Custom Level", id);
 				return;
 			}
 			String author = JsonHelper.getString(json, "author", "level.author.unknown");
@@ -160,7 +158,7 @@ public class LevelDataManager extends JsonDataLoader
 		});
 		setLevels(builtinBuilder.build(), true);
 		setLevels(customBuilder.build(), false);
-		Ultracraft.LOGGER.info("Loaded " + levels.size() + " Builtin Levels and " + customLevels.size() + " Custom Levels");
+		Ultracraft.LOGGER.info("Loaded {} Builtin Levels and {} Custom Levels", levels.size(), customLevels.size());
 	}
 	
 	public static LevelData getLevelData(Identifier id)
@@ -171,7 +169,7 @@ public class LevelDataManager extends JsonDataLoader
 			return customLevels.get(id);
 		else
 		{
-			Ultracraft.LOGGER.warn("Couldn't find Level '" + id + "' in loaded lists!");
+			Ultracraft.LOGGER.warn("Couldn't find Level '{}' in loaded lists!", id);
 			return ERR_DATA;
 		}
 	}
@@ -204,7 +202,7 @@ public class LevelDataManager extends JsonDataLoader
 	
 	public static boolean isCustomLevelsPresent()
 	{
-		return customLevels.size() > 0;
+		return !customLevels.isEmpty();
 	}
 	
 	public static void sync(ServerPlayerEntity player)
@@ -212,6 +210,8 @@ public class LevelDataManager extends JsonDataLoader
 		PacketByteBuf buf = new PacketByteBuf(Unpooled.buffer());
 		buf.writeCollection(levels.entrySet(), LevelData::serialize);
 		buf.writeCollection(customLevels.entrySet(), LevelData::serialize);
+		buf.writeCollection(LevelCollectionManager.layers.entrySet(), LevelCollection::serialize);
+		buf.writeCollection(LevelCollectionManager.customLayers.entrySet(), LevelCollection::serialize);
 		ServerPlayNetworking.send(player, PacketRegistry.SEND_LEVELS_PACKET_ID, buf);
 	}
 }